move saveUnmatched button and function to PredictedDog screen
